move tests to lib
